+++ CHANGELOG.md
@@ -35,6 +35,7 @@
   * GandiV5.domain -> GandiV5::Domain.fetch
   * GandiV5.mailboxes -> GandiV5::Email::Mailbox.list
   * GandiV5.mailbox_slots -> GandiV5::Email::Slot.list
+  * GandiV5::LiveDNS.domains -> GandiV5::LiveDNS::Domain.list
 
 ## Version 0.1.0